requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
exa-py
pydantic
python-dotenv
requests
